Actualizacion de autorizacion para categorias

+++ app/Http/Controllers/Api/Admin/CategorieController.php
@@ -3,24 +3,23 @@
 namespace App\Http\Controllers\Api\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\CategorieResource;
 use App\Models\Categorie;
 use Illuminate\Http\Request;
 
 class CategorieController extends Controller
 {
-    public function __construct(){
-        $this->middleware('can:manage-category');
-    }
-
     //Funcion para mostrar todas las categorias registradas en la base de datos
     public function index()
     {
+        $this->authorize('index', Categorie::class);
+
         $categories = Categorie::all();
         return $this->sendResponse(
             message: "Categories returned successfully",
             code: 200,
             result: [
-                'categories' => $categories,
+                'categories' => CategorieResource::collection($categories),
             ]
         );
     }
@@ -28,11 +27,13 @@ class CategorieController extends Controller
     //Funcion para mostrar una categoria en especifico
     public function show(Categorie $categorie)
     {
+        $this->authorize('show', Categorie::class);
+
         return $this->sendResponse(
             message: "Category returned successfully",
             code: 200,
             result: [
-                'categorie' => $categorie,
+                'categorie' => new CategorieResource($categorie),
             ]
         );
     }
@@ -40,10 +41,12 @@ class CategorieController extends Controller
     //Funcion para crear una categoria en la base de datos
     public function store(Request $request)
     {
+        $this->authorize('store', Categorie::class);
+
         //Se valida la informacion de la categoria
         $request->validate([
             'name' => 'required|unique:categories|max:255',
-            'imagen' => 'required'
+            'imagen' => 'required|string'
         ]);
 
         //Se crea la categoria
@@ -54,7 +57,7 @@ class CategorieController extends Controller
             message: "Category created successfully",
             code: 201,
             result: [
-                'categorie' => $categorie,
+                'categorie' => new CategorieResource($categorie),
             ]
         );
     }
@@ -62,21 +65,21 @@ class CategorieController extends Controller
     //Funcion para actualizar una categoria en la base de datos
     public function update(Request $request, Categorie $categorie)
     {
+        $this->authorize('update', Categorie::class);
         //Se valida la informacion de la categoria
         $request->validate([
-            'name' => 'required|unique:categories,name,'.$categorie->id.'|max:255',
+            'name' => 'required|max:255',
             'imagen' => 'required'
         ]);
 
         //Se actualiza la categoria
         $categorie->update($request->all());
-
         //Se retorna la categoria actualizada
         return $this->sendResponse(
             message: "Category updated successfully",
             code: 200,
             result: [
-                'categorie' => $categorie,
+                'categorie' => new CategorieResource($categorie),
             ]
         );
     }
@@ -84,6 +87,7 @@ class CategorieController extends Controller
     //Funcion para eliminar una categoria en la base de datos
     public function destroy(Categorie $categorie)
     {
+        $this->authorize('delete', Categorie::class);
         //Se elimina la categoria
         $categorie->delete();
